gear_core_backend::error

Trait BackendAllocSyscallError

pub trait BackendAllocSyscallError: Sized {
    type ExtError: BackendSyscallError;

    // Required method
    fn into_backend_error(self) -> Result<Self::ExtError, Self>;
}

A trait for conversion of the externalities memory management error to api error.

If the conversion fails, then Self is returned in the Err variant.

Required Associated Types

type ExtError: BackendSyscallError

Required Methods

fn into_backend_error(self) -> Result<Self::ExtError, Self>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.